requirements.txt
anyio==3.4.0
fastapi==0.70.1
greenlet==1.1.2
idna==3.3
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pydantic==1.9.0
python-decouple==3.5
sniffio==1.2.0
SQLAlchemy==1.4.29
starlette==0.16.0
typing-extensions==4.0.1
